Move some functionality from FluxMaps to FluxEstimate
