+++ gammapy/estimators/flux_map.py
@@ -6,38 +6,12 @@ from astropy.table import Table
 from astropy.utils import classproperty
 from gammapy.data import GTI
 from gammapy.maps import MapCoord, Map
-from gammapy.estimators.core import FluxEstimate, OPTIONAL_QUANTITIES_COMMON
+from gammapy.estimators.core import FluxEstimate, OPTIONAL_QUANTITIES_COMMON, REQUIRED_MAPS, OPTIONAL_MAPS
 from gammapy.estimators.flux_point import FluxPoints
 from gammapy.modeling.models import SkyModel, Models
 from gammapy.utils.scripts import make_path
 
 __all__ = ["FluxMaps"]
-
-
-REQUIRED_MAPS = {
-    "dnde": ["dnde"],
-    "e2dnde": ["e2dnde"],
-    "flux": ["flux"],
-    "eflux": ["eflux"],
-    "likelihood": ["norm"],
-}
-
-
-# TODO: add an entry for is_ul?
-OPTIONAL_MAPS = {
-    "dnde": ["dnde_err", "dnde_errp", "dnde_errn", "dnde_ul"],
-    "e2dnde": ["e2dnde_err", "e2dnde_errp", "e2dnde_errn", "e2dnde_ul"],
-    "flux": ["flux_err", "flux_errp", "flux_errn", "flux_ul"],
-    "eflux": ["eflux_err", "eflux_errp", "eflux_errn", "eflux_ul"],
-    "likelihood": [
-        "norm_err",
-        "norm_errn",
-        "norm_errp",
-        "norm_ul",
-        "norm_scan",
-        "stat_scan",
-    ],
-}
 
 
 log = logging.getLogger(__name__)
@@ -310,20 +284,6 @@ class FluxMaps(FluxEstimate):
         return cls.from_dict(
             maps=maps, sed_type=sed_type, reference_model=reference_model, gti=gti
         )
-
-    @staticmethod
-    def _validate_type(maps, sed_type):
-        """Check that map input is valid and correspond to one of the SED type."""
-        try:
-            required = set(REQUIRED_MAPS[sed_type])
-        except KeyError:
-            raise ValueError(f"Unknown SED type.")
-
-        if not required.issubset(maps.keys()):
-            missing = required.difference(maps.keys())
-            raise ValueError(
-                "Missing maps for sed type '{}':" " {}".format(sed_type, missing)
-            )
 
     @classmethod
     def from_dict(cls, maps, sed_type="likelihood", reference_model=None, gti=None):
